requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0
scikit-learn>=1.1.0
xgboost>=1.6.0
lightgbm>=3.3.0
torch>=1.12.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.8.0
fastapi>=0.100.0
uvicorn>=0.18.0
pydantic>=2.0.0
requests>=2.28.0
aiohttp>=3.8.0
beautifulsoup4>=4.11.0
selenium>=4.5.0
nfl-data-py>=0.3.0
click>=8.0.0
pyyaml>=6.0
python-dotenv>=0.20.0
schedule>=1.2.0
tqdm>=4.64.0
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.971